requirements.txt
annotated-types==0.6.0
anyio==3.7.1
beautifulsoup4==4.12.2
cachetools==5.3.2
camelot-py==0.11.0
certifi==2023.11.17
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.0
cryptography==41.0.5
cycler==0.12.1
et-xmlfile==1.1.0
fastapi==0.104.1
fonttools==4.43.1
google-api-core==2.14.0
google-api-python-client==2.108.0
google-auth==2.23.4
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
googleapis-common-protos==1.61.0
h11==0.14.0
httplib2==0.22.0
idna==3.4
kiwisolver==1.4.5
matplotlib==3.8.0
numpy==1.26.2
oauthlib==3.2.2
opencv-python-headless==4.8.1.78
openpyxl==3.1.2
packaging==23.2
pandas==2.1.3
pdfminer.six==20221105
Pillow==10.0.1
protobuf==4.25.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.5.1
pydantic_core==2.14.3
pyparsing==3.1.1
pypdf==3.17.1
PyPDF2==3.0.1
PyPDF4==1.27.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
starlette==0.27.0
tabulate==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.1.0
uvicorn==0.24.0.post1